+++ src/containerapp/azext_containerapp/_params.py
@@ -462,10 +462,10 @@ def load_arguments(self, _):
         c.argument('polling_interval', type=int, help="Interval to check each event source in seconds. Defaults to 30s.", default=30)
 
     with self.argument_context('containerapp job create') as c:
-        c.argument('system_assigned', help='System assigned identity.', action='store_true')
+        c.argument('system_assigned', help='Boolean indicating whether to assign system-assigned identity.', action='store_true')
         c.argument('trigger_type', help='Trigger type. Schedule | Event | Manual')
-        c.argument('user_assigned', help='User assigned identity.')
-
+        c.argument('user_assigned', nargs='+', help='Space-separated user identities to be assigned.')
+ 
     with self.argument_context('containerapp job', arg_group='Scale') as c:
         c.argument('min_executions', type=int, help="Minimum number of job executions to run per polling interval.")
         c.argument('max_executions', type=int, help="Maximum number of job executions to run per polling interval.")